Written questions from Idris lectures
These are (summarised) questions from the first two Idris lectures, and I think it'd be good to publish these and fill in the answers. For some of them, it's probably best that I answer myself, but it might also be useful for others to contribute to a lot of them, so maybe it's worth including them here for some collaborative editing.

I don't know the right place to put this file - maybe better in a subdirectory?